+++ com/denvys5/uraniumswordmod/machines/uraniumduplicator/DuplicatorRecipes.java
@@ -99,6 +99,6 @@ public class DuplicatorRecipes{
         }
         while (!this.func_151397_a(itemstack, (ItemStack)entry.getKey()));
 
-        return ((int)entry.getValue());
+        return ((Integer)entry.getValue()).intValue();
     }
 }
